update the release notes and tests pkg
