+++ util/sdkutil/string.go
@@ -2,7 +2,6 @@ package sdkutil
 
 import (
 	"fmt"
-	"math/rand"
 	"strings"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
@@ -30,14 +29,4 @@ func FormatDec(d sdk.Dec) string {
 // e.g. 4.000uumee -> 4 uumee and 3.500ibc/abcd -> 3.5 ibc/abcd
 func FormatDecCoin(c sdk.DecCoin) string {
 	return fmt.Sprintf("%s %s", FormatDec(c.Amount), c.Denom)
-}
-
-func GenerateString(length uint) string {
-	// character set used for generating a random string in GenerateString
-	charset := "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789"
-	bytes := make([]byte, length)
-	for i := range bytes {
-		bytes[i] = charset[rand.Intn(len(charset))] //nolint
-	}
-	return string(bytes)
 }
